e.timeStamp is broken in opera 11.64
In opera 11.64 on OSX the timeStamp attribute is always --2147483648.
This prevent the mouse based events from firing and stops scrolling
with the mouse from working after the first event. In order to fix this
a different method of timestamp generation is required. The alternative
I went with was new Date().getTime().