Use forceZoom instead of always setting zoom
Seems the library relied on the side affect of the zoom not always showing and messed with max/min scroll positions. This just adds the additional flag we needed and doesn't affect the underlying library's use case. This is rather hacky and I'm sure there's a better way to do it.